+++ src/pages/NoticesPage/NoticeCategoryItem/components/LearnMore.jsx
@@ -1,0 +1,26 @@
+import React, { useState } from 'react';
+import Button from 'shared/components/Button';
+import Icon from 'shared/components/Icon/Icon';
+
+const LearnMore = ({ onButtonClick }) => {
+  const [isHoveredOrFocused, setIsHoveredOrFocused] = useState(false);
+  return (
+    <Button
+      style={{
+        width: '100%',
+        marginTop: '20px',
+        transition: 'stroke 250ms ease',
+      }}
+      onMouseEnter={() => setIsHoveredOrFocused(true)}
+      onMouseLeave={() => setIsHoveredOrFocused(false)}
+      onFocus={() => setIsHoveredOrFocused(true)}
+      onBlur={() => setIsHoveredOrFocused(false)}
+      onClick={onButtonClick}
+    >
+      Learn More
+      {isHoveredOrFocused && <Icon id="paw" f="#FEF9F9" w="24" h="24" />}
+    </Button>
+  );
+};
+
+export default LearnMore;
