--- src/pages/NoticesPage/NoticeModalTest/NoticeModalTest.jsx
@@ -1,51 +1,0 @@
-import React from 'react';
-import { createPortal } from 'react-dom';
-
-const modalRoot = document.getElementById('modal-root');
-
-const ModalNoticeTest = ({ close }) => {
-  const closeModal = ({ target, currentTarget, code }) => {
-    if (target === currentTarget || code === 'Escape') {
-      close();
-    }
-  };
-
-  React.useEffect(() => {
-    document.addEventListener('keydown', closeModal);
-    return () => {
-      document.removeEventListener('keydown', closeModal);
-    };
-  });
-
-  return createPortal(
-    <div
-      onClick={closeModal}
-      style={{
-        zIndex: '1999',
-        position: 'fixed',
-        top: 0,
-        left: 0,
-        right: 0,
-        bottom: 0,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        backgroundColor: 'rgba(0, 0, 0, 0.5)',
-      }}
-    >
-      <div
-        onClick={e => e.stopPropagation()}
-        style={{
-          backgroundColor: 'white',
-          padding: '1em',
-          borderRadius: '10px',
-        }}
-      >
-        <h2>Modal</h2>
-      </div>
-    </div>,
-    modalRoot
-  );
-};
-
-export default ModalNoticeTest;
